CoreEntity oluşturuldu, app.java boşaltıldı,
