+++ .vscode/settings.json
@@ -1,0 +1,3 @@
+{
+    "java.home" : "C:/Program Files/Java/jdk1.8.0_191"
+}
