+++ core.base/src/main/java/tr/edu/izu/yam/core/base/IEntity.java
@@ -1,0 +1,5 @@
+package tr.edu.izu.yam.core.base;
+
+public interface IEntity {
+
+}
